create sirius class
